refactor heatmap utils into their own module
